Complete TaskList and Storage classes
