+++ src/main/java/luigi/Parser.java
@@ -1,0 +1,9 @@
+package luigi;
+
+import luigi.tasks.Task;
+
+import java.util.Scanner;
+
+public class Parser {
+
+}
