+++ src/main/java/luigi/Ui.java
@@ -1,0 +1,15 @@
+package luigi;
+
+import java.util.Scanner;
+
+public class Ui {
+    private final Scanner scanner;
+
+    public Ui() {
+        this.scanner = new Scanner(System.in);
+    }
+
+    public String readCommand() {
+        return scanner.nextLine().trim();
+    }
+}
