+++ src/main/java/luigi/tasks/Event.java
@@ -1,4 +1,4 @@
-package tasks;
+package luigi.tasks;
 
 import java.time.LocalDateTime;
 import java.time.LocalDate;
