+++ src/main/java/luigi/tasks/Task.java
@@ -1,4 +1,4 @@
-package tasks;
+package luigi.tasks;
 public class Task {
     protected String description;
     protected boolean isDone;
